Accept keysym/unicode symbols on keys
This is an enhancement to make non-ASCII symbol keys usable for
implementation and configuration of Awesome.

The enhancement was proposed and had been developed under the
initiative of Daniel Hahler.  Thanks to his sharing his results with our
community, we gradually deepened our understanding on the issue.  This
commit is the fruit his enthusiasm beared.