Fix invalid memory usage in event.c
p_dup really only duplicates the given region of memory. It does not append a
zero byte. Thus, the string we are using here was not zero-terminated.

Fix this by just using lua_pushlstring() so that we do not have to worry about a
null terminating the string at all.

Signed-off-by: Uli Schlachter <psychon@znc.in>