awesomerc.lua: swap meaning of j and k keys
Using j for previous and k for next is more intuitive.
When using more than 2 screens, mod+ctrl+j was focusing the screen
to the right of the current one, while mod+ctrl+k was focusing the
one to the left.(though j is located to the left of k).