screen: detect monitor configuration change
Xorg server from git[1] generates ConfigNotify event when monitor
configuration changes.

We handle the event, but only trigger required awesome_restart() if
screen size is changed. It's not what happens in case of monitor
configuration change.

Let's relax the condition and call awesome_restart() in response to all
ConfigNotify events to root window.

I'm not entirely sure that it wouldn't lead to superfluous restarts, but
I don't see an option.

[1] dbe8d03c42f0 ("randr: Send ConfigNotify when manual monitor list changes")

Signed-off-by: Kirill A. Shutemov <kirill@shutemov.name>