xresources theme: Use a RecordingSurface for wallpaper
This theme draws a minimal wallpaper directly. This commit replaces the
ImageSurface of the wallpaper with a RecordingSurface. This shouldn't have any
visible effect, except for reducing memory usage, because an ImageSurface needs
a lot of memory and is only eventually released by the garbage collector. A
RecordingSurface instead just records the operations that were done which needs
a lot less memory for the simple operations done here.

Signed-off-by: Uli Schlachter <psychon@znc.in>